+++ docs/.vitepress/config.ts
@@ -40,11 +40,45 @@ export default defineConfig({
             '/': [
                 {
                     text: 'Guide',
-                    link: '/guide',
+                    children: [
+                        {
+                            text: 'Intro',
+                            link: '/intro',
+                        },
+                    ],
+                },
+                {
+                    text: 'Recipes',
+                    children: [
+                        {
+                            text: 'Generate a Key Pair',
+                            link: '/recipes/key-pair-gen',
+                        },
+                        {
+                            text: 'Initialize a Client',
+                            link: '/recipes/init-client',
+                        },
+                        {
+                            text: 'Query: Find All Assets',
+                            link: '/recipes/query-all-assets',
+                        },
+                        {
+                            text: 'Transaction: Register Asset Definition',
+                            link: '/recipes/tx-reg-asset-definition',
+                        },
+                        {
+                            text: 'Events: Listen for 5 Commited Blocks',
+                            link: '/recipes/events-blocks',
+                        },
+                    ],
                 },
                 {
                     text: 'API',
                     link: '/api/',
+                },
+                {
+                    text: 'To Do',
+                    link: '/todo',
                 },
             ],
         },