+++ packages/crypto/scripts/build-wasm.ts
@@ -1,7 +1,6 @@
 import consola from 'consola';
-import del from 'del';
 import makeDir from 'make-dir';
-import { $, cd, path } from 'zx';
+import { $, cd, path, glob } from 'zx';
 
 const CRATE_ROOT_DIR = path.resolve(__dirname, '../crypto-rs/iroha_crypto_wasm');
 const BUILD_TMP_DIR = path.resolve(CRATE_ROOT_DIR, '.tmp-pkg');
@@ -11,8 +10,11 @@ function computeDistForTarget(targetName: string): string {
     return path.resolve(__dirname, `../packages/target-${targetName}`);
 }
 
-function necessaryArtifacts(outName: string): string[] {
-    return [`${outName}_bg*`, `${outName}.*`];
+async function necessaryArtifacts(outName: string): Promise<string[]> {
+    const items = await Promise.all(
+        [`${outName}_bg*`, `${outName}.*`].map((x) => path.join(BUILD_TMP_DIR, x)).map((x) => glob(x)),
+    );
+    return items.flat();
 }
 
 export async function build_wasm() {
@@ -31,18 +33,17 @@ export async function build_wasm() {
         consola.info('Building', config.target);
 
         cd(CRATE_ROOT_DIR);
-        await $`wasm-pack build \
-            --target ${config.target}
-            --out-dir ${BUILD_TMP_DIR}
+        await $`wasm-pack build \\
+            --target ${config.target} \\
+            --out-dir ${BUILD_TMP_DIR} \\
             --out-name ${BUILD_OUT_NAME}
         `;
 
         const dist = config.distDir;
         await makeDir(dist);
-        await del([path.join(dist, '*')]);
 
         cd(BUILD_TMP_DIR);
-        await $`cp ${necessaryArtifacts(BUILD_OUT_NAME)} ${dist}`;
+        await $`cp ${await necessaryArtifacts(BUILD_OUT_NAME)} ${dist}`;
     }
 
     consola.success('Done!');