+++ packages/data-model/gulpfile.ts
@@ -1,2 +1,7 @@
-export { default as compile_json } from './scripts/compile-json';
-export { default as generate } from './scripts/generate';
+import { series } from 'gulp';
+import compile_json from './scripts/compile-json';
+import generate from './scripts/generate';
+
+export const gen = series(compile_json, generate);
+
+export { compile_json, generate };