+++ packages/i64-fixnum/CHANGELOG.md
@@ -1,5 +1,16 @@
 # @iroha2/i64-fixnum
 
+## 0.3.0
+
+### Minor Changes
+
+-   92c5a9a: Rename functions to make their purpose clearer; update doccomments
+
+### Patch Changes
+
+-   33a58c5: Bump `@scale-codec/util` version
+-   92c5a9a: Add library top-level short documentation
+
 ## 0.2.0
 
 ### Minor Changes